+++ lib/src/repositories/daemon/daemon_rpc_methods_extension.dart
@@ -365,7 +365,6 @@ extension DaemonRpcMethodsExtension on DaemonClient {
   /// Retrieve all peers connected
   Future<GetPeersResult> getPeers() async {
     final result = await sendRequest(DaemonMethod.getPeers);
-    print(result);
     return GetPeersResult.fromJson(result as Map<String, dynamic>);
   }